backend/digital_lawyer/requirements.txt
Django==5.0.1
psycopg2-binary==2.9.9
djangorestframework==3.14.0
djoser==2.2.2
django-filter==23.5
django-cors-headers==4.3.1
django-storages==1.14.2
boto3==1.34.30
shortuuid==1.0.11

# DEV
drf-spectacular==0.27.1
drf-spectacular-sidecar==2024.1.1